refactor: minor code improvements
