+++ src/ni_measurement_plugin_packager/_support/_package_info.py
